requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
faster-whisper>=1.0.0
pyannote.audio>=3.1.1
numpy>=1.24.0
scipy>=1.11.0
